+++ tests/Console/Command/installer/invalid_config.php
@@ -1,0 +1,3 @@
+<?php
+
+$sugar_config_si['test'] = 'foo';